pyproject.toml
[tool.poetry]
name = "webdev"
version = "0.1.0"
description = ""
authors = ["Rishabh Chopde <rishikumarikuf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
Flask-SQLAlchemy = "^2.5.1"
psycopg2 = "^2.9.1"
google-auth-oauthlib = "^0.4.6"
python-dotenv = "^0.19.2"
Authlib = "^0.15.5"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
